Update.
2000-09-07  Ulrich Drepper  <drepper@redhat.com>

	* assert/__assert.c: Undefine NDEBUG before include <assert.h>.

	* sysdeps/unix/sysv/linux/msgsnd.c (msgsnd): Add cast to avoid warning.
	* sysdeps/unix/sysv/linux/shmat.c (shmat): Likewise.

+++ sysdeps/unix/sysv/linux/shmat.c
@@ -48,7 +48,7 @@ shmat (shmid, shmaddr, shmflg)
 #endif
 
   result = (void *__unbounded) INLINE_SYSCALL (ipc, 5, IPCOP_shmat, shmid, shmflg,
-					       __ptrvalue (&raddr),
+					       (long int) __ptrvalue (&raddr),
 					       __ptrvalue (shmaddr));
   if ((unsigned long) result <= -(unsigned long) SHMLBA)
     result = raddr;
